+++ iscc_sdk/options.py
@@ -86,6 +86,11 @@ class SdkOptions(BaseSettings):
         description="ISCC_SDK_VIDEO_FPS - Frames per second to process for video hash (ignored when 0).",
     )
 
+    video_store_mp7sig: bool = Field(
+        False,
+        description="ISCC_SDK_VIDEO_STORE_MP7SIG - Store extracted MP7 Video as <videofile>.iscc.mp7sig",
+    )
+
     @validator("image_max_pixels")
     def set_pillow(cls, v):
         Image.MAX_IMAGE_PIXELS = v