+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.5.7 - Unreleased
 - Updated dependencies
+- Fixed mkdocstrings
 
 ## 0.5.6 - 2023-04-28
 - Don´t install tika more than once per session