+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.5.7 - Unreleased
 - Added parallel processing of ISCC-UNITs
 - Handle video thumbnail extraction errors gracefully
+- Add basic command line interface
 - Updated dependencies
 - Fixed mkdocstrings